+++ .gitignore
@@ -1,1 +1,7 @@
 /.idea
+env
+ingestion/src/config/__pycache__/
+ingestion/src/consumer/__pycache__/
+ingestion/src/processor/__pycache__/
+ingestion/src/uploader/__pycache__/
+ingestion/src/writer/__pycache__/
