+++ docker-compose.yml
@@ -103,8 +103,37 @@ services:
     volumes:
       - ./minio_lake:/data
     command: server /data --console-address ":9001"
-    
+
+  clickhouse:
+    image: yandex/clickhouse-server:latest
+    container_name: clickhouse
+    ports:
+      - "9010:9010"
+      - "8123:8123"
+      - "9009:9009"
+    environment:
+      - CLICKHOUSE_DB=test_db
+    volumes:
+      - clickhouse_data:/var/lib/clickhouse
+    networks:
+      - subnet
+    depends_on:
+      - minio
+
+  clickhouse-client:
+    image: yandex/clickhouse-client:latest
+    container_name: clickhouse-client
+    command: ["clickhouse-client", "--host", "clickhouse"]
+    depends_on:
+      - clickhouse
+    networks:
+      - subnet
+
 networks:
   subnet:
     name: subnet
     driver: bridge
+
+volumes:
+  clickhouse_data:
+    driver: local
